# env-rest-openapi-specifications/v1/openapi-environments-v1.yaml
openapi: 3.0.3
info:
  title: Api Documentation
  description: Api Documentation
  version: "1.1"
  contact:
    name: OurCompany
    url: https://www.ourcompany.com/company/about-ourcompany/contact-us.html
servers:
  - url: 'http://localhost:8080'
    description: Inferred Url
...
paths:
...
  '/api/environments/{environmentId}':
    get:
      tags:
        - environment-controller
      summary: getEnvironment
      operationId: getEnvironment
      parameters:
        - name: environmentId
          in: path
          description: environmentId
          required: true
          style: simple
          schema:
            type: string
            format: uuid
        - name: full
          in: query
          description: full
          required: false
          style: form
          schema:
            type: boolean

      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/EnvironmentFullVer1View'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
        - environment-controller
      summary: createSystem
      operationId: createSystem
      parameters:
        - name: environmentId
          in: path
          description: environmentId
          required: true
          style: simple
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSystemDto'
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/System'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - environment-controller
      summary: delete
      operationId: deleteEnvironment
      parameters:
        - name: environmentId
          in: path
          description: environmentId
          required: true
          style: simple
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
components:
  schemas:
    Connection:
      title: Connection
      type: object
      properties:
        connectionType:
          type: string
        created:
          type: integer
          format: int64
        createdBy:
          type: string
          format: uuid
        description:
          type: string
        id:
          type: string
          format: uuid
        modified:
          type: integer
          format: int64
        modifiedBy:
          type: string
          format: uuid
        name:
          type: string
        parameters:
          type: object
          additionalProperties:
            type: string
        services:
          type: array
          items:
            type: string
        sourceTemplateId:
          type: string
          format: uuid
        systemId:
          type: string
          format: uuid
    ConnectionDto:
      title: ConnectionDto
      type: object
      properties:
        connectionType:
          type: string
        created:
          type: integer
          format: int64
        createdBy:
          type: string
          format: uuid
        description:
          type: string
        id:
          type: string
          format: uuid
        modified:
          type: integer
          format: int64
        modifiedBy:
          type: string
          format: uuid
        name:
          type: string
        parameters:
          type: object
          additionalProperties:
            type: string
        projectId:
          type: string
          format: uuid
        services:
          type: array
          items:
            type: string
        sourceTemplateId:
          type: string
          format: uuid
        systemId:
          type: string
          format: uuid
    ConnectionFullVer1View:
      title: ConnectionFullVer1View
      type: object
      properties:
        connectionType:
          type: string
        created:
          type: integer
          format: int64
        createdBy:
          type: string
          format: uuid
        description:
          type: string
        id:
          type: string
          format: uuid
        modified:
          type: integer
          format: int64
        modifiedBy:
          type: string
          format: uuid
        name:
          type: string
        parameters:
          type: object
          additionalProperties:
            type: string
        services:
          type: array
          items:
            type: string
        sourceTemplateId:
          type: string
          format: uuid
        systemId:
          type: string
          format: uuid
...
    CreateSystemDto:
      title: CreateSystemDto
      type: object
      properties:
        connections:
          type: array
          items:
            type: object    #$ref: '#/components/schemas/ConnectionDto'
        dateOfCheckVersion:
          type: integer
          format: int64
        dateOfLastCheck:
          type: integer
          format: int64
        description:
          type: string
        environmentId:
          type: string
          format: uuid
        externalName:
          type: string
        externalId:
          type: string
          format: uuid
        id:
          type: string
          format: uuid
        linkToSystemId:
          type: string
          format: uuid
        mergeByName:
          type: boolean
        name:
          type: string
        parametersGettingVersion:
          $ref: '#/components/schemas/ParametersGettingVersion'
        parentSystemId:
          type: string
          format: uuid
        serverITF:
          type: object #$ref: '#/components/schemas/ServerItf'
        status:
          type: string
          enum:
            - FAIL
            - NOTHING
            - PASS
            - WARN
        systemCategoryId:
          type: string
          format: uuid
        version:
          type: string
    Environment:
      title: Environment
      type: object
      properties:
        created:
          type: integer
          format: int64
        createdBy:
          type: string
          format: uuid
        description:
          type: string
        graylogName:
          type: string
        id:
          type: string
          format: uuid
        modified:
          type: integer
          format: int64
        modifiedBy:
          type: string
          format: uuid
        name:
          type: string
        projectId:
          type: string
          format: uuid
        systems:
          type: array
          items:
            type: object #$ref: '#/components/schemas/System'
        ssmSolutionAlias:
          type: string
        ssmInstanceAlias:
          type: string
        consulEgressConfigPath:
          type: string
        tags:
          type: array
          items:
            type: string
...
    EnvironmentFullVer1View:
      title: EnvironmentFullVer1View
      type: object
      properties:
        created:
          type: integer
          format: int64
        createdBy:
          type: string
          format: uuid
        description:
          type: string
        graylogName:
          type: string
        id:
          type: string
          format: uuid
        modified:
          type: integer
          format: int64
        modifiedBy:
          type: string
          format: uuid
        name:
          type: string
        projectId:
          type: string
          format: uuid
        systems:
          type: array
          items:
            type: object #$ref: '#/components/schemas/SystemFullVer1View'
        ssmSolutionAlias:
          type: string
        ssmInstanceAlias:
          type: string
        consulEgressConfigPath:
          type: string
        tags:
          type: array
          items:
            type: string
...
    ParametersGettingVersion:
      title: ParametersGettingVersion
      type: object
      properties:
        headers:
          type: string
        parameters:
          type: string
        parsingType:
          type: string
          enum:
            - JSONPATH
            - NONE
            - REGEXP
        parsingValue:
          type: string
        type:
          type: string
          enum:
            - BY_HTTP_ENDPOINT
            - BY_HTTP_ENDPOINT_BASIC_AUTH
            - BY_KUBERNETES_CONFIGMAP
            - BY_KUBERNETES_IMAGES
            - BY_OPENSHIFT_CONFIGURATION
            - BY_SHELL_SCRIPT
            - BY_SQL_QUERY
            - BY_SSM
    ParametersGettingVersionFullVer1View:
      title: ParametersGettingVersionFullVer1View
      type: object
...
    ServerItf:
      title: ServerItf
      type: object
      properties:
        name:
          type: string
        url:
          type: string
    ServerItfFullVer1View:
      title: ServerItfFullVer1View
      type: object
...
    System:
      title: System
      type: object
      properties:
        connections:
          type: array
          items:
            type: object #$ref: '#/components/schemas/Connection'
        created:
          type: integer
          format: int64
        createdBy:
          type: string
          format: uuid
        dateOfCheckVersion:
          type: integer
          format: int64
        dateOfLastCheck:
          type: integer
          format: int64
        description:
          type: string
        environmentIds:
          type: array
          items:
            type: object #$ref: '#/components/schemas/Environment'
        environments:
          type: array
          items:
            type: object #$ref: '#/components/schemas/Environment'
        externalName:
          type: string
        externalId:
          type: string
          format: uuid
        id:
          type: string
          format: uuid
        linkToSystemId:
          type: string
          format: uuid
        mergeByName:
          type: boolean
        modified:
          type: integer
          format: int64
        modifiedBy:
          type: string
          format: uuid
        name:
          type: string
        parametersGettingVersion:
          $ref: '#/components/schemas/ParametersGettingVersion'
        parentSystemId:
          type: string
          format: uuid
        serverITF:
          type: object #$ref: '#/components/schemas/ServerItf'
        status:
          type: string
          enum:
            - FAIL
            - NOTHING
            - PASS
            - WARN
        systemCategory:
          type: object #$ref: '#/components/schemas/SystemCategory'
        systemCategoryId:
          type: string
          format: uuid
        version:
          type: string
...
    SystemCategory:
      title: SystemCategory
      type: object
      properties:
        created:
          type: integer
          format: int64
        createdBy:
          type: string
          format: uuid
        description:
          type: string
        id:
          type: string
          format: uuid
        modified:
          type: integer
          format: int64
        modifiedBy:
          type: string
          format: uuid
        name:
          type: string
...
    SystemFullVer1View:
      title: SystemFullVer1View
      type: object
      properties:
        connections:
          type: array
          items:
            type: object #$ref: '#/components/schemas/ConnectionFullVer1View'
        created:
          type: integer
          format: int64
        createdBy:
          type: string
          format: uuid
        dateOfCheckVersion:
          type: integer
          format: int64
        dateOfLastCheck:
          type: integer
          format: int64
        description:
          type: string
        environmentIds:
          type: array
          items:
            type: object #$ref: '#/components/schemas/EnvironmentFullVer1View'
        externalName:
          type: string
        externalId:
          type: string
          format: uuid
        id:
          type: string
          format: uuid
        linkToSystemId:
          type: string
          format: uuid
        mergeByName:
          type: boolean
        modified:
          type: integer
          format: int64
        modifiedBy:
          type: string
          format: uuid
        name:
          type: string
        parametersGettingVersion:
          $ref: '#/components/schemas/ParametersGettingVersionFullVer1View'
        parentSystemId:
          type: string
          format: uuid
        serverITF:
          type: object #$ref: '#/components/schemas/ServerItfFullVer1View'
        status:
          type: string
          enum:
            - FAIL
            - NOTHING
            - PASS
            - WARN
        systemCategoryId:
          type: string
          format: uuid
        version:
          type: string
...